requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.0.0
Werkzeug==0.11.2
argparse==1.2.1
itsdangerous==0.24
numpy==1.10.1
scipy==0.16.1
selenium==2.48.0
wsgiref==0.1.2
